Update integration tests
- Adjustments to how integration tests are set up and ran
