+++ .circleci/config.yml
@@ -20,6 +20,9 @@ jobs:
       - run: curl -L https://codeclimate.com/downloads/test-reporter/test-reporter-latest-linux-amd64 > ./cc-test-reporter
       - run: chmod +x ./cc-test-reporter
       - run: ./cc-test-reporter before-build
+      - run:
+          name: Setup testing database
+          command: npm run init:test-db
       - run: 
           name: Runnning tests
           command: npm run cover
