+++ package.json
@@ -105,6 +105,7 @@
     "preinstall": "",
     "postinstall": "npm run uswds",
     "init": "node ./tools/init.js",
+    "init:test-db": "node ./test/database/setup.js",
     "demo": "node ./tools/demo.js",
     "build": "gulp build",
     "migrate": "echo 'Migrate commands exist for the following actions [:up, :down, :check]'",
@@ -115,9 +116,9 @@
     "start": "npm run migrate:up && node app.js",
     "test": "npm run test:unit",
     "test:unit": "node ./node_modules/mocha/bin/mocha test/bootstrap.test.js test/unit/setup.js test/unit/*.test.js test/unit/**/*.test.js --opts test/test/mocha.opts",
-    "test:integration": "node ./node_modules/mocha/bin/mocha test/bootstrap.test.js test/integration/services/setup.js test/integration/services/*.test.js",
+    "test:integration": "node ./node_modules/mocha/bin/mocha test/bootstrap.test.js test/integration/services/*.test.js",
     "jshint": "find assets/js/backbone -name \\*js -exec ./node_modules/jshint/bin/jshint {} \\;",
-    "cover": "nyc --reporter=lcov --reporter=text ./node_modules/mocha/bin/mocha test/bootstrap.test.js test/unit/setup.js test/unit/*.test.js test/unit/**/*.test.js --opts test/test/mocha.opts test/integration/services/setup.js test/integration/services/*.test.js",
+    "cover": "nyc --reporter=lcov --reporter=text ./node_modules/mocha/bin/mocha test/bootstrap.test.js test/unit/setup.js test/unit/*.test.js test/unit/**/*.test.js --opts test/test/mocha.opts test/integration/services/*.test.js",
     "cover:unit": "nyc --reporter=lcov --reporter=text npm run test:unit",
     "cover:integration": "nyc --reporter=lcov --reporter=text -x \"api/model/**\" -x \"utils/**\" -x \"test/**\" npm run test:integration",
     "debug": "node debug app.js",
